chore: Extract interval value to a separate file
